--- index.html
@@ -1,37 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-  <meta charset="UTF-8">
-  <meta name="viewport" content="width=device-width, initial-scale=1.0">
-  <link rel="stylesheet" href="https://stackpath.bootstrapcdn.com/bootstrap/4.5.0/css/bootstrap.min.css" integrity="sha384-9aIt2nRpC12Uk9gS9baDl411NQApFmC26EwAOH8WgZl5MYYxFfc+NcPb1dKGj7Sk" crossorigin="anonymous">
-  <link rel="stylesheet" href="./css/main.css">
-  <title>Tic-Tac-Toe</title>
-</head>
-<body>
-  <div class="container">
-    <div class="message">
-
-    </div>
-    <div class="d-flex justify-content-center align-items-center flex-column game">
-      <form class="m-4" id='form-start'>
-        <div class="form-group">
-          <label for="player-1">Player 1 name:</label>
-          <input type="text" class="form-control" id="player-1" name="player-1">
-        </div>
-        <div class="form-group">
-          <label for="player-2">Player 2 name:</label>
-          <input type="text" class="form-control" id="player-2" name="player-2">
-        </div>
-        <button type="submit" id="start" class="btn btn-primary m-1">Start Playing</button>
-      </form>
-    </div>
-
-  </div>
-
-  <script src="https://code.jquery.com/jquery-3.5.1.slim.min.js" integrity="sha384-DfXdz2htPH0lsSSs5nCTpuj/zy4C+OGpamoFVy38MVBnE+IbbVYUew+OrCXaRkfj" crossorigin="anonymous"></script>
-  <script src="https://cdn.jsdelivr.net/npm/popper.js@1.16.0/dist/umd/popper.min.js" integrity="sha384-Q6E9RHvbIyZFJoft+2mJbHaEWldlvI9IOYy5n3zV9zzTtmI3UksdQRVvoxMfooAo" crossorigin="anonymous"></script>
-  <script src="https://stackpath.bootstrapcdn.com/bootstrap/4.5.0/js/bootstrap.min.js" integrity="sha384-OgVRvuATP1z7JjHLkuOU7Xw704+h835Lr+6QL9UvYjZE3Ipu6Tp75j7Bh/kR0JKI" crossorigin="anonymous"></script>
-  <script src="./js/logic.js" ></script>
-  <script src="./js/index.js" ></script>
-</body>
-</html>
